chore(test): organize test folders
