+++ .prettierignore
@@ -1,3 +1,3 @@
 # Tests files
-src/generators/api-links/test/fixtures/
+src/generators/api-links/__tests__/fixtures/
 *.snapshot
