+++ src/generators/api-links/__tests__/fixtures.test.mjs.snapshot
