+++ src/generators/api-links/__tests__/fixtures/buffer.js
