+++ src/generators/api-links/__tests__/fixtures/class.js
