+++ src/generators/api-links/__tests__/fixtures/mod.js
@@ -2,10 +2,8 @@
 
 // A module may export one or more methods.
 
-function foo() {
-}
-
+function foo() {}
 
 module.exports = {
-  foo
+  foo,
 };
