+++ src/generators/api-links/__tests__/fixtures/prototype.js
