+++ src/linter/__tests__/fixtures/descriptors.mjs
