+++ src/linter/reporters/__tests__/github.test.mjs
@@ -1,8 +1,12 @@
-import assert from 'node:assert';
+import assert from 'node:assert/strict';
 import { describe, it } from 'node:test';
 
-import github from '../../reporters/github.mjs';
-import { errorIssue, infoIssue, warnIssue } from '../fixtures/issues.mjs';
+import {
+  errorIssue,
+  infoIssue,
+  warnIssue,
+} from '../../__tests__/fixtures/issues.mjs';
+import github from '../github.mjs';
 
 describe('github', () => {
   it('should write to stdout with the correct fn based on the issue level', t => {
@@ -12,13 +16,13 @@ describe('github', () => {
     github(warnIssue);
     github(errorIssue);
 
-    assert.strictEqual(process.stdout.write.mock.callCount(), 3);
+    assert.equal(process.stdout.write.mock.callCount(), 3);
 
     const callsArgs = process.stdout.write.mock.calls.map(call =>
       call.arguments[0].trim()
     );
 
-    assert.deepStrictEqual(callsArgs, [
+    assert.deepEqual(callsArgs, [
       '::notice file=doc/api/test.md,line=1,endLine=1::This is a INFO issue',
       '::warning file=doc/api/test.md,line=1,endLine=1::This is a WARN issue',
       '::error file=doc/api/test.md,line=1,endLine=1::This is a ERROR issue',
