+++ src/linter/rules/__tests__/duplicate-stability-nodes.test.mjs
@@ -2,7 +2,7 @@ import { deepStrictEqual, strictEqual } from 'node:assert';
 import { describe, it, mock } from 'node:test';
 
 import { LINT_MESSAGES } from '../../constants.mjs';
-import { duplicateStabilityNodes } from '../../rules/duplicate-stability-nodes.mjs';
+import { duplicateStabilityNodes } from '../duplicate-stability-nodes.mjs';
 
 // Mock data structure for creating test entries
 const createStabilityNode = (value, line = 1, column = 1) => ({
