+++ src/linter/rules/__tests__/fixtures/invalid-change-version-subprocess.mjs
@@ -3,7 +3,7 @@ import { mock } from 'node:test';
 
 import dedent from 'dedent';
 
-import { invalidChangeVersion } from '../../rules/invalid-change-version.mjs';
+import { invalidChangeVersion } from '../../invalid-change-version.mjs';
 
 const yamlContent = dedent`
 <!-- YAML
