+++ src/utils/queries/__tests__/index.test.mjs
@@ -1,7 +1,7 @@
 import { strictEqual, deepStrictEqual } from 'node:assert';
 import { describe, it } from 'node:test';
 
-import createQueries from '../queries/index.mjs';
+import createQueries from '../index.mjs';
 
 describe('createQueries', () => {
   it('should add YAML metadata correctly', () => {
